refactor(client): reorganize protocol services
