+++ Cargo.toml
@@ -105,7 +105,7 @@ tokio-btls = "0.5.6"
 tokio = { version = "1.52.3", default-features = false }
 compio = { version = "0.19.2", features = ["io", "io-compat"], optional = true }
 wreq-rt = { version = "0.2.2-rc.2", default-features = false }
-wreq-proto = { version = "0.2.3", default-features = false }
+wreq-proto = { version = "0.2.6", default-features = false }
 tower = { version = "0.5.3", default-features = false, features = [
     "timeout",
     "util",
