+++ src/client/pool/cache.rs
