+++ src/client/pool/map.rs
